resolve: move resolving double-star globs to fs package
Move the functions to resolve double star glob patterns from the resolve to the
fs package.
This will allow to reuse the functionality in other parts of the code.
